Move: IAM without options (test)
- Move the test that uses IAM without options to the `Default`
  test(s).
